LambdaRecomendedRAG/requirements.txt
openai
python-dotenv
transformers
torch
PyPDF2
supabase
boto3
google-cloud-vision
python-bidi==0.4.2
llama-index
faiss-cpu==1.8.0
fpdf2
python-bidi
numpy==1.26.4
llama-cloud==0.0.17
llama-index==0.11.7
llama-index-agent-openai==0.3.1
llama-index-cli==0.3.0
llama-index-core==0.11.7
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.3
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.1
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-faiss==0.2.1
llama-parse==0.5.2
langchain
langchain-community
langchain_openai
rank_bm25
PyMuPDF
selenium